requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.16
requests==2.32.3
pydantic==2.9.2
openai==1.3.0
python-dotenv==1.0.0
